Small fixes.

+++ www/js/controllers/settingsCtrl.js
@@ -255,7 +255,7 @@
         $scope.deleteGroup = function () {
             $cordovaDialogs.prompt(deleteGroupMsg, deleteGroupTitle , ['OK', 'Cancel'], '')
                 .then(function (result) {
-                    if (result.input1.length > 0 && result.input1 == $scope.groupTitle) {
+                    if (result && result.input1 !== 'undefined' && result.input1.length > 0 && result.input1 == $scope.groupTitle) {
                         console.log('Deleting group ' + result.input1);
                         gdApi.deleteGroup(Number($stateParams.gid)).then(
                             function (groupDeleted) {
